Updated the path flattening code to correctly observe tolerance values, and made the tolerances adapt to the scaling being used when drawing to a transformed graphics context. Fixed a small issue with focus listeners, and a mac menu dismissal problem.

+++ src/gui/graphics/geometry/juce_Line.h
@@ -209,7 +209,8 @@ public:
                                               ValueType perpendicularDistance) const throw()
     {
         const Point<ValueType> delta (end - start);
-        const double length = juce_hypot (delta.getX(), delta.getY());
+        const double length = juce_hypot ((double) delta.getX(),
+                                          (double) delta.getY());
         if (length == 0)
             return start;
